requirements.txt
fastapi
uvicorn[standard]
python-docx
pdfplumber
pypdf
faiss-cpu
numpy
tqdm
rich
requests
